+++ test/gurka/gurka.cc
@@ -97,7 +97,6 @@ std::string build_valhalla_request(const std::string& location_type,
 
   rapidjson::Value co(rapidjson::kObjectType);
   if (!custom_speed_types) {
-    rapidjson::Value dt(rapidjson::kObjectType);
     rapidjson::Value speed_types(rapidjson::kArrayType);
     speed_types.PushBack("freeflow", allocator);
     speed_types.PushBack("constrained", allocator);
@@ -601,11 +600,9 @@ valhalla::Api do_action(const valhalla::Options::Action& action,
       break;
     case valhalla::Options::expansion:
       json_str = actor.expansion(request_json, nullptr, &api);
-      std::cout << json_str << std::endl;
       break;
     case valhalla::Options::isochrone:
       json_str = actor.isochrone(request_json, nullptr, &api);
-      std::cout << json_str << std::endl;
       break;
     default:
       throw std::logic_error("Unsupported action");